cilium: create lib for encryption
To allow both bpf_network and bpf_host to use the same code add a encrypt.h
file to put common routines.
